Switched back from open4 to IO.popen.
open4 was added as a dependency when I was trying to parse STDERR and needed to work with STDIN. As it turns out, $stdin is good enough.  Redirecting STDERR to STDOUT seems to be good enough as well.  While open4 is probably still a cleaner approach, its reliance on fork() means it doesn't work on non-JRuby in Windows (echnically it doesn't work in JRuby either, but JRuby has an IO.popen4 special method that does virtually the same thing and we were using that instead).